Added null input check and made optimization
I wasn't really sure *what* to do for the null check, so clarification is appreciated.
Optimization: A bit of the program started with an empty array `sentences` and iterated over `wordsList` to fill `sentences` with the same value over and over again. This was not so good because the program had to find new memory to expand `sentences` over and over again. 
Now, it gets a big chunk of memory all at once and fills it with one value using the convenient builtin `Array#fill`.